move WeatherConditionIcons to util  + update imports
